Swapping ex3p_o1.cpp for ex3_o1.cpp
